refactor: move reproduce() method to abstract class to eliminate duplication
