+++ src/main/java/org/example/model/Rabbit.java
@@ -11,23 +11,4 @@ public class Rabbit extends Herbivore {
     public Rabbit() {
         super(7, 0.45, 2, 150, 2);
     }
-
-    @Override
-    public void reproduce(Cell cell, StatisticTracker tracker) throws NoSuchMethodException, InvocationTargetException, InstantiationException, IllegalAccessException {
-        for (Animal animal : new ArrayList<>(cell.getAnimals())) {
-            if (animal instanceof Rabbit && animal != this) {
-                if (new ArrayList<>(cell.getAnimals()).stream()
-                        .filter(a -> a instanceof Rabbit)
-                        .toList().size() < maxAmount) {
-                    if(RandomUtil.getRandomBoolean(1, 4)) {
-                        cell.addAnimal(this.getClass().getConstructor().newInstance());
-                        tracker.increment(this.getClass().getSimpleName() + AnimalType.valueOf(this.getClass().getSimpleName().toUpperCase()).getEmoji(), "reproduced");
-//                        System.out.println("Rabbit reproduced");
-                    }
-                    return;
-                }
-            }
-        }
-    }
-
 }
